+++ R/rc_input_event_amendment.R
@@ -7,7 +7,7 @@
 #'
 #' @details This function increases temporal detail for time series of C inputs of organic amendments.
 #' The inputs for organic amendments are organised in the data.table amendment, where the carbon inputs has the unit kg C / ha.
-#' dt. time Must contain integerish columns \code{year} (>= 0) and \code{month} (1..12), numeric column \code{time}, 
+#' dt. time must contain integerish columns \code{year} (>= 0) and \code{month} (1..12), numeric column \code{time}, 
 #' and unique \code{(year, month)} pairs over the simulation window.
 #'
 #' The output is an EVENT object.